+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.2
 
 - Fixes default `max_size` parameter being `0` instead of `1`
+
+## 1.0.3
+
+- Adds autoscaling id, arn name outputs
+- Adds `cluster_launch_configuration_id` output